geocode addr vectors

Addresses are matched to a set of reference addresses; if unmatched, they are next matched to TIGER street ranges by number and street name, returning the centroid of the (unionized) matched street range(s).

Usage

addr_match_geocode(
  x,
  ref_addr = cagis_addr()$cagis_addr,
  ref_s2,
  county = "39061",
  year = "2022"
)

Arguments

x

an addr vector (or character vector of address strings) to geocode

ref_addr

an addr vector to search for matches in

ref_s2

a s2_cell vector of locations for each ref_addr

county

character county identifer for TIGER street range files to search for matches in

year

character year for TIGER street range files to search for matches in

Value

a tibble with columns: addr contains x converted to an addr vector, s2 contains the resulting geocoded s2 cells as an s2cell vector, match_method is a factor with levels ref_addr, tiger_range, none to record the method by which each address was matched

Details

Performance was compared to the degauss geocoder (see /inst/compare_geocoding_to_degauss.R) using real-world addresses in voter_addresses(). Match success rates were highly similar:

addr_matched	degauss_matched	n	perc

TRUE	TRUE	224974	92.9%

FALSE	TRUE	13713	5.7%

FALSE	FALSE	2687	1.1%

TRUE	FALSE	759	0.3%

Among those that were geocoded by both, 97.7% were geocoded to the same census tract:

ct_agree	bg_agree	n	s2_dist_ptiles (5th, 25th, 50th, 75th, 95th)	perc

TRUE	TRUE	217176	14.7, 24.3, 39, 68.9, 153.6	96.5%

FALSE	FALSE	5067	21.9, 40.8, 212.5, 6616.5, 20222.2	2.3%

TRUE	FALSE	2731	19.6, 28.6, 41.2, 95.1, 569.8	1.2%

Examples

cagis_s2 <-
  cagis_addr()$cagis_addr_data |>
  purrr::modify_if(\(.) length(.) > 0 && nrow(.) > 1, dplyr::slice_sample, n = 1) |>
  purrr::map_vec(purrr::pluck, "cagis_s2", .default = NA, .ptype = s2::s2_cell())
addr_match_geocode(x = sample(voter_addresses(), 100), ref_s2 = cagis_s2)
#> matching addresses in 22/38 ZIP codes [3.6s elapsed] 
#> matching addresses in 35/38 ZIP codes [6.6s elapsed] 
#> matching addresses in 38/38 ZIP codes [7s elapsed] 
#> 
#> # A tibble: 100 × 3
#>                                          addr s2               match_method
#>                                        <addr> <s2cell>         <fct>       
#>  1       9756 Ross Avenue Cincinnati OH 45242 88405386f89c2569 ref_addr    
#>  2        1621 1st Avenue Cincinnati OH 45205 8841b5dbee7c2bbf ref_addr    
#>  3      9854 Arvin Avenue Cincinnati OH 45231 88404bbd8470be7d ref_addr    
#>  4   7324 Hillside Avenue Cincinnati OH 45233 8841cc0783e2b755 ref_addr    
#>  5   10885 Carolitrace Road Harrison OH 45030 NA               none        
#>  6       2944 Hull Avenue Cincinnati OH 45211 8841b5783ed3f5e9 ref_addr    
#>  7 5310 Timberchase Court Cincinnati OH 45247 8841cab7436adf0d ref_addr    
#>  8 5236 Valley Ridge Road Cincinnati OH 45247 8841cada7af313b9 ref_addr    
#>  9        10650 Marvin Road Harrison OH 45030 88402dc45e31f31f ref_addr    
#> 10   3700 Grovedale Place Cincinnati OH 45208 8841b2632a33ccc1 ref_addr    
#> # ℹ 90 more rows