Match an addr vector to TIGER street ranges

Match an addr vector to TIGER street ranges

Usage

addr_match_tiger_street_ranges(
  x,
  county = "39061",
  year = "2022",
  street_only_match = TRUE,
  summarize = c("none", "union", "centroid")
)

Arguments

x

an addr vector to match

county

character string of county identifier

year

year of tigris product

street_only_match

logical; consider an addr that matches on street name, but does not have a street number within the listed ranges a match?

summarize

optionally summarize matched street ranges as their union or centroid

Value

a list of matched tigris street range tibbles; a NULL value indicates that no street name was matched; if street_only_match is FALSE, a street range tibble with zero rows indicates that although a street was matched, there was no range containing the street number

Examples

my_addr <- as_addr(c("224 Woolper Ave", "3333 Burnet Ave", "33333 Burnet Ave", "609 Walnut St"))

addr_match_tiger_street_ranges(my_addr, county = "39061", street_only_match = FALSE)
#> $`224 Woolper Avenue`
#> # A tibble: 1 × 4
#>   TLID      s2_geography                                              from    to
#>   <chr>     <s2_geography>                                           <dbl> <dbl>
#> 1 103924294 LINESTRING (-84.511197 39.149684, -84.511752 39.149788,…   100   299
#> 
#> $`3333 Burnet Avenue`
#> # A tibble: 2 × 4
#>   TLID      s2_geography                                            from    to
#>   <chr>     <s2_geography>                                         <dbl> <dbl>
#> 1 103925697 LINESTRING (-84.500403 39.14089, -84.500289 39.141892)  3301  3399
#> 2 103925699 LINESTRING (-84.500525 39.139737, -84.500403 39.14089)  3247  3398
#> 
#> $`33333 Burnet Avenue`
#> # A tibble: 0 × 4
#> # ℹ 4 variables: TLID <chr>, s2_geography <s2_geography>, from <dbl>, to <dbl>
#> 
#> $`609 Walnut Street`
#> NULL
#> 

addr_match_tiger_street_ranges(my_addr, county = "39061", summarize = "centroid")
#> $`224 Woolper Avenue`
#> # A tibble: 1 × 4
#>   TLID      s2_geography                    from    to
#>   <chr>     <s2_geography>                 <dbl> <dbl>
#> 1 103924294 POINT (-84.5148163 39.1499943)   100   299
#> 
#> $`3333 Burnet Avenue`
#> # A tibble: 1 × 4
#>   TLID                s2_geography                    from    to
#>   <chr>               <s2_geography>                 <dbl> <dbl>
#> 1 103925697-103925699 POINT (-84.5004091 39.1408146)  3247  3399
#> 
#> $`33333 Burnet Avenue`
#> # A tibble: 1 × 4
#>   TLID                                                  s2_geography  from    to
#>   <chr>                                                 <s2_geograp> <dbl> <dbl>
#> 1 103925448-103925451-103925453-103925455-103925697-10… POINT (-84.…  2000  3599
#> 
#> $`609 Walnut Street`
#> NULL
#> 

addr_match_tiger_street_ranges(my_addr, county = "39061", summarize = "centroid") |>
  dplyr::bind_rows() |>
  dplyr::mutate(census_bg_id = s2_join_tiger_bg(s2::as_s2_cell(s2_geography)))
#> # A tibble: 3 × 5
#>   TLID                                     s2_geography  from    to census_bg_id
#>   <chr>                                    <s2_geograp> <dbl> <dbl> <chr>       
#> 1 103924294                                POINT (-84.…   100   299 390610070002
#> 2 103925697-103925699                      POINT (-84.…  3247  3399 390610270002
#> 3 103925448-103925451-103925453-103925455… POINT (-84.…  2000  3599 390610270002