CAGIS Addresses

CAGIS Addresses

Usage

cagis_addr()

Value

An example tibble created from the CAGIS addresses with a pre-calculated, unique cagis_addr vector column. The cagis_addr_data column is a list of tibbles because one CAGIS address can correspond to multiple parcel identifiers and address-level data (place, type, s2, etc.). See data-raw/make_cagis_addr.R for source code to create data, including filtering criteria:

use only addresses that have STATUS of ASSIGNED or USING and are not orphaned (ORPHANFLG == "N")

omit addresses with ADDRTYPEs that are milemarkers (MM), parks (PAR), infrastructure projects (PRJ), cell towers (CTW), vacant or commercial lots (LOT), and other miscellaneous non-residential addresses (MIS, RR, TBA)

s2 cell is derived from LONGITUDE and LATITUDE fields in CAGIS address database

Examples

cagis_addr()
#> # A tibble: 307,206 × 2
#>                                   cagis_addr    cagis_addr_data
#>                                       <addr> <list<tibble[,6]>>
#>  1            1 1st Street Addyston OH 45001            [1 × 6]
#>  2   1 A Tanglewood Lane Cincinnati OH 45224            [1 × 6]
#>  3        1 Abbey Place Indian Hill OH 45243            [1 × 6]
#>  4          1 Abbey Hill North Bend OH 45052            [1 × 6]
#>  5    1 Abbington Ridge Indian Hill OH 45242            [1 × 6]
#>  6         1 Adelle Walk Greenhills OH 45218            [1 × 6]
#>  7         1 Albert Place Mariemont OH 45227            [1 × 6]
#>  8         1 Alcott Lane Greenhills OH 45218            [1 × 6]
#>  9 1 Anderson Ferry Road Cincinnati OH 45204            [1 × 6]
#> 10        1 Andover Road Greenhills OH 45218            [1 × 6]
#> # ℹ 307,196 more rows