Get tigris street range geography files from census.gov

Downloaded files are cached in tools::R_user_dir("addr", "cache"). Street ranges with missing minimum or maximum address numbers are excluded.

Usage

get_tiger_street_ranges(county, year = "2022")

Arguments

county

character string of county identifier

year

year of tigris product

Value

a list of tibbles, one for each street name, with TLID, s2_geography, from, and to columns

Examples

Sys.setenv("R_USER_CACHE_DIR" = tempfile())
get_tiger_street_ranges("39061")[1001:1004]
#> <list_of<
#>   tbl_df<
#>     TLID        : character
#>     s2_geography: s2_geography
#>     from        : double
#>     to          : double
#>   >
#> >[4]>
#> $`Bridgecreek Dr`
#> # A tibble: 5 × 4
#>   TLID      s2_geography                                              from    to
#>   <chr>     <s2_geography>                                           <dbl> <dbl>
#> 1 627893439 LINESTRING (-84.518038 39.237373, -84.518037 39.237248)   9301  9327
#> 2 627893440 LINESTRING (-84.518038 39.237373, -84.51802 39.237481, …  9310  9326
#> 3 637586321 LINESTRING (-84.517673 39.236089, -84.517688 39.236225,…  9200  9294
#> 4 637586322 LINESTRING (-84.517948 39.236947, -84.517967 39.237009,…  9296  9306
#> 5 645615840 LINESTRING (-84.51794 39.237697, -84.517789 39.237925, …  9328  9399
#> 
#> $`Bridgeknoll Ct`
#> # A tibble: 1 × 4
#>   TLID      s2_geography                                              from    to
#>   <chr>     <s2_geography>                                           <dbl> <dbl>
#> 1 103952990 LINESTRING (-84.645545 39.144769, -84.645801 39.144859,…  3200  3299
#> 
#> $`Bridgeport Ln`
#> # A tibble: 3 × 4
#>   TLID      s2_geography                                              from    to
#>   <chr>     <s2_geography>                                           <dbl> <dbl>
#> 1 103943261 LINESTRING (-84.512157 39.30345, -84.512079 39.304385)   12100 12199
#> 2 103943264 LINESTRING (-84.512233 39.302483, -84.512181 39.303229,… 12046 12099
#> 3 103943265 LINESTRING (-84.512337 39.301232, -84.512233 39.302483)  11900 12045
#> 
#> $`Bridges Rd`
#> # A tibble: 3 × 4
#>   TLID      s2_geography                                              from    to
#>   <chr>     <s2_geography>                                           <dbl> <dbl>
#> 1 103940363 LINESTRING (-84.364851 39.093519, -84.365029 39.093423,…  6900  7106
#> 2 103956809 LINESTRING (-84.363794 39.090174, -84.362162 39.090046)   7101  7145
#> 3 103956810 LINESTRING (-84.362162 39.090046, -84.362018 39.090026,…  7138  7299
#>