Example National Address Database addresses

A small baked example fixture containing Hamilton County, Ohio National Address Database (NAD version 22L) rows needed by package examples and tests. See ?nad for more information about the National Address Database.

The fixture was derived once from the Hamilton County NAD source data and is not intended to be regenerated by package users. Use nad() when you need complete county data.

Usage

nad_example_data(match_prepared = FALSE)

Arguments

match_prepared

logical; return the example data preprocessed with addr_match_prepare()?

Value

If match_prepared = FALSE, a tibble with 5,000 rows and 7 columns. If match_prepared = TRUE, an addr_match_index.

Examples

nad_example_data()
#> # A tibble: 5,000 × 7
#>    nad_addr            subaddress uuid  date_update s2    address_type parcel_id
#>    <addr>              <chr>      <chr> <date>      <s2_> <chr>        <chr>    
#>  1 3359 QUEEN CITY Av… NA         {E3A… 2025-03-30  -6.7… Unknown      NA       
#>  2 1040 KREIS Ln CINC… NA         {D05… 2025-03-30  -6.7… Unknown      NA       
#>  3 9960 DALY Rd CINCI… NA         {109… 2025-03-30  -6.1… Unknown      NA       
#>  4 413 VOLKERT Pl CIN… NA         {1BB… 2025-03-30  -6.7… Unknown      NA       
#>  5 8519 LINDERWOOD Ln… NA         {F78… 2025-03-30  -6.6… Unknown      NA       
#>  6 6361 BEECHMONT Ave… NA         {6D4… 2025-03-30  -6.6… Unknown      NA       
#>  7 10466 ADVENTURE Ln… NA         {1D1… 2025-03-30  -6.1… Unknown      NA       
#>  8 3156 LOOKOUT Cir C… NA         {AE8… 2025-03-30  -6.6… Unknown      NA       
#>  9 310 WYOMING Ave CI… NA         {331… 2025-03-30  -6.1… Unknown      NA       
#> 10 118 SPRINGFIELD Pi… NA         {F3E… 2025-03-30  -6.1… Unknown      NA       
#> # ℹ 4,990 more rows
nad_example_data(match_prepared = TRUE)
#> # An addr_match_index object with 4,823 addrs in 60 ZIP code groups:
#> 45238: 87
#> 45205: 89
#> 45231: 51
#> 45219: 17
#> 45255: 90
#> 45230: 87
#> 45242: 87
#> 45208: 2
#> 45215: 84
#> 45239: 87
#> 45229: 90
#> 45207: 8
#> 45237: 34
#> 45240: 87
#> 45224: 76
#> 45211: 87
#> 45227: 71
#> 45243: 89
#> 45246: 55
#> 45251: 89
#> 45206: 73
#> 45202: 91
#> 45030: 92
#> 45002: 89
#> 45209: 89
#> 45140: 88
#> 45244: 90
#> 45236: 86
#> 45220: 87
#> 45232: 87
#> 45247: 92
#> 45233: 90
#> 45212: 80
#> 45213: 87
#> 45248: 90
#> 45241: 85
#> 45221: 74
#> 45249: 90
#> 45214: 86
#> 45001: 91
#> 45013: 94
#> 45014: 92
#> 45033: 88
#> 45041: 88
#> 45051: 88
#> 45052: 26
#> 45111: 93
#> 45147: 90
#> 45150: 90
#> 45174: 87
#> 45203: 89
#> 45204: 90
#> 45216: 89
#> 45217: 89
#> 45218: 89
#> 45223: 88
#> 45225: 87
#> 45226: 88
#> 45228: 86
#> 45252: 88
#> .