Get names for tiger street ranges

TIGER primary feature names are read from compressed feature-name databases for each county and Census vintage. If not already present, compressed addrfeat (address feature) shapefiles are downloaded from the Census FTP site to the addr user data directory.

When reading into R, the data is filtered to addressable MTFCCs (S1100, S1200, S1400, S1640) that have a name.

Usage

tiger_feat_names(county, year, redownload = FALSE)

Arguments

county

character string of county FIPS identifier

year

character year of the Census TIGER/Line product

redownload

logical, length 1; re-download the cached TIGER ZIP file?

Value

a tibble with unique LINEARID and addr columns

Examples

tiger_feat_names("39061", "2025")
#> # A tibble: 11,640 × 2
#>    LINEARID      addr_street       
#>    <chr>         <addr_str>        
#>  1 110480468364  Quarry St         
#>  2 110480484495  Hill and Dale Dr  
#>  3 110480461502  W Old 8th St      
#>  4 110480469966  Old Chimney Ct    
#>  5 110480463772  Old Court St      
#>  6 110480475686  Old Red Bank Rd   
#>  7 1103676806008 Old Paddys Run Rd 
#>  8 110480471931  Old Town Ct       
#>  9 1106092774309 Old Remington Rd  
#> 10 110480482631  Old Indian Hill Rd
#> # ℹ 11,630 more rows